Update app.py
modified and updated code to incorporate different types of quires into flask application this includes an UPDATE query and GET and PUT methods. when a PUT request is recived it updates the information of the pets specifies by its id using an update query. the SELECT query fetches all the pets from the 'pet_info' table